requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.36
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
# Updated to fix Docker cache issue
# Note: Using Gmail SMTP (smtplib is part of Python standard library, no package needed)